EPP-Configuration-Generator/requirements.txt
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PyYAML==6.0.2
